requirements.txt
# python utils
requests==2.24.0
openpyxl==3.0.5

Django==3.1.3

# django utils
dj_database_url==0.5.0
django-storages==1.7.2
pik-django-utils==2.0.5

# HISTORY
django-simple-history==2.12.0

# db
redis==3.5.3
psycopg2==2.8.6
django-redis==4.12.1

# celery
celery==5.0.2

# sentry
sentry-sdk==0.19.2

# prod
uwsgi==2.0.18

# dev
django_debug_toolbar==3.1.1
django-extensions==3.0.9
werkzeug==1.0.1

# tests
factory_boy==2.12.0
freezegun==0.3.12
pdbpp==0.10.0
pytest==6.1.2
pytest-mock==3.3.1
pytest-cov==2.10.1
pytest-django==4.1.0
pytest-base-url==1.4.2
pytest-benchmark==3.2.3
# tests/style
prospector==1.3.1
Faker==4.14.2

# metrics/monitoring
django-health-check==3.16.1
